pyproject.toml
[tool.poetry]
name = "eeg_research"
version = "0.1.0"
description = "All tools and pipelines developed for eeg research"
authors = ["Samuel Louviot <sam.louvito@gmail.com>"]
license = "LGPL-3.0"
readme = "README.md"
packages = [{include = "eeg_research", from = "src"}]

[tool.poetry.dependencies]
python = "~3.12"
simple-term-menu = "^1.6.4"
pybids = "^0.16.4"
mne = "^1.7.0"
numpy = "^1.26.4"
scipy = "^1.13.0"
matplotlib = "^3.8.4"
asrpy = "^0.0.3"
pyprep = "^0.4.3"
neurokit2 = "^0.2.7"
pybv = "^0.7.5"
eeglabio = "^0.0.2.post4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
mypy = "^1.10.0"
pre-commit = "^3.7.0"
pytest-cov = "^5.0.0"
ruff = "^0.4.2"
ipykernel = "^6.29.4"
pytest-mock = "^3.14.0"

[tool.poetry.group.docs.dependencies]
pdoc = "^14.4.0"

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
testpaths = [
  "tests"
]

[tool.mypy]
ignore_missing_imports = true
exclude = "^data/"

[tool.ruff]
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv"
]
line-length = 88
indent-width = 4
src = ["src"]
target-version = "py311"
extend-exclude = [
  "*.ipynb",
  "src/eeg_research/labs"
]

[tool.ruff.lint]
select = ["ANN", "D", "E", "F", "I"]
ignore = [
  "ANN101",  # self should not be annotated.
  "ANN102"  # cls should not be annotated.
]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"
